refactor: 코드 스타일 적용 테스트
